Disable mangling of filenames for page templates and ZCML definitions when rendering the registration list.  The zcml file was never rendered on this page and the page template paths were only visible as tooltips.
The mangling algorithm involves attempting to import various subpaths until a valid one is found, which over the course of 116-odd registrations is a lot of calls to __import__.  The unmangled version is the absolute path, rather than the plone.app.customerize.tests/foo.pt format.

This change improves performance of registrations.html by 4000% and leaves the customize page unchanged.

svn path=/plone.app.customerize/trunk/; revision=32021